# src/main/java/com/sphereon/ms/eidas/api/model/signature/CadesSignatureVerifyRequest.java
/*
 * Copyright (C) 2022 Sphereon BV
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.sphereon.ms.eidas.api.model.signature;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonProperty;

public class CadesSignatureVerifyRequest {
    private final String signature;

    @JsonCreator
    public CadesSignatureVerifyRequest(@JsonProperty("signature") String signature) {
        this.signature = signature;
    }

    public String getSignature() {
        return signature;
    }
}